+++ quickbooks/objects/taxrate.py
@@ -1,9 +1,11 @@
 from six import python_2_unicode_compatible
-from .base import QuickbooksManagedObject, QuickbooksTransactionEntity, Ref
+
+from quickbooks.mixins import ListMixin, ReadMixin
+from .base import QuickbooksTransactionEntity, Ref, QuickbooksBaseObject
 
 
 @python_2_unicode_compatible
-class TaxRate(QuickbooksManagedObject, QuickbooksTransactionEntity):
+class TaxRate(QuickbooksTransactionEntity, QuickbooksBaseObject, ReadMixin, ListMixin):
     """
     QBO definition: A TaxRate object represents rate applied to calculate tax liability. Use the TaxService
     entity to create a taxrate.
@@ -17,6 +19,8 @@ class TaxRate(QuickbooksManagedObject, QuickbooksTransactionEntity):
 
     def __init__(self):
         super(TaxRate, self).__init__()
+
+        # All values are readonly - TaxRates cannot be created or modified with the TaxRate object
         self.Name = ""
         self.Description = ""
         self.RateValue = 0