+++ CHANGELOG.rst
@@ -15,6 +15,7 @@ Changelog
     * Fixed issues with default values on Term.
     * Fixed issues with default values on TaxService and TaxRateDetails.
     * Fixed issues that prevented save from working on TaxService.
+    * Removed unsupported save method from TaxRate.
 
 * 0.4.0 (June 15, 2016)
     * Added a way of disconnecting a Quickbooks Account to client.